requirements.txt
asgiref==3.5.2
boto3==1.24.39
botocore==1.27.39
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jmespath==1.0.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
